Factor out item-specific code.
